eesp.org AAD changes.
this commit is still a work in progress. It need polishing.
How to show what would be covered in AAD, also with  Crypt Offset.

Also further clarifaction on all crypto suites that use integerity
protection would be covered in this update.

This would also need a change to IANA IKEv2 registry add the column
EESP there. Then do we need to upate EACH RFCs?